requirements.txt
# BigMoeHunter Requirements - Lightweight Version
# No external AI API keys required - uses built-in rule-based system

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.25.2
folium==0.15.0
geopandas==0.14.1
shapely==2.0.2
beautifulsoup4==4.12.2
lxml==4.9.3
matplotlib==3.8.2
seaborn==0.13.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Optional AI libraries (uncomment if you want to use local AI models)
# transformers==4.35.2
# torch==2.1.1
# sentence-transformers==2.2.2

# Development Dependencies
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
